+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [0.3.9] - 2020-11-10
 ### Added
 - New command: `slice-img` exports an image from a 2D slice. Unlike `export`
   this supports N-dimensional volumes and arbitrary slicing dimensions.